test/api/devicehub_client/pyproject.toml
[tool.poetry]
name = "devicehub_client"
version = "1.3.10"
description = "A client library for accessing DeviceHub"
authors = []
readme = "README.md"
packages = [
    {include = "devicehub_client"},
]
include = ["CHANGELOG.md", "devicehub_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
